+++ src/pages/editionCv/EditExperience.js
@@ -1,20 +1,26 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import {NotificationManager} from 'react-notifications';
+import { useNavigate } from "react-router-dom";
 
 export default function EditExperience({data}) {
     const [experience, setExperience] = useState('');
     const [societe, setSociete] = useState('');
     const [dateExp, setDateExp] = useState('');
+    const navigate = useNavigate();
     const [descriptionExp, setDescriptionExp] = useState('');
     const [datefin, setDateFin] = useState('');
     useEffect(() => {
+       getExperiences();
+    }, [data])
+
+    const getExperiences = () => {
         axios.get(`experiences/${data}`).then(resp => {
             if(resp.status === 200) {
                 setExperience(resp.data)
             }
         }).catch(error => console.log(error))
-    }, [data])
+    }
     
     const editExperience = () => {
         const formdata = new FormData;
@@ -26,6 +32,7 @@ export default function EditExperience({data}) {
             axios.put(`experiences/${data}`, formdata).then(resp => {
                 if(resp.status === 200) {
                     NotificationManager.success('Modifier avec succée', 'Modification valdataer', 4000);
+                    getExperiences();
                 }
             }).catch(error => console.log(error))
         }else {
@@ -34,12 +41,21 @@ export default function EditExperience({data}) {
     }
 
     const DeleteExperience = () => {
-        axios.delete(`experiences/${data}`).then(resp => {
-            if(resp.status === 204) {
-                NotificationManager.success('Vous avec supprimer une experience', 'Suprimer avec succèes', 4000);
-            }
-        }).catch(error => console.log(error));
+        if(window.confirm("Vous êtes sûr ?") === true){
+            axios.delete(`experiences/${data}`).then(resp => {
+                if(resp.status === 204) {
+                    NotificationManager.success('Vous avec supprimer une experience', 'Suprimer avec succèes', 4000);
+                    navigate('/cv');
+                }
+            }).catch(error => console.log(error));
+        }
     }
+    
+    const dexp = [];
+    for (let years = 1990; years <= 2040; years++) {
+        dexp.push(`${years}`);
+    }
+
     return (
         <form>
             <div className="row mb-3">
@@ -51,6 +67,7 @@ export default function EditExperience({data}) {
                         type="text" className="form-control" defaultValue={experience.societe}
                         onChange={(e) => setSociete(e.target.value)}
                     />
+
                 </div>
             </div>
             <div className="row mb-3">
@@ -58,16 +75,26 @@ export default function EditExperience({data}) {
                     <h6 className="mb-0">Date d'experience</h6>
                 </div>
                 <div className="col-sm-4 text-secondary">
-                    <input 
-                        type="date" className="form-control" defaultValue={experience.datexp}
+                    <select 
+                        type="text" className="form-control" defaultValue={experience.datexp}
                         onChange={(e) => setDateExp(e.target.value)}
-                    />
+                    >
+                        <option defaultChecked >{experience.datexp}</option>
+                        {dexp.map((dates) => (
+                            <option value={dates} key={dates}>{dates}</option>
+                        ))}
+                    </select>
                 </div>
                 <div className="col-sm-5 text-secondary">
-                    <input 
+                    <select 
                         type="date" className="form-control" defaultValue={experience.datefin}
                         onChange={(e) => setDateFin(e.target.value)}
-                    />
+                    >
+                        <option defaultChecked >{experience.datefin}</option>
+                         {dexp.map((dates) => (
+                            <option value={dates} key={dates}>{dates}</option>
+                        ))}
+                    </select>
                 </div>
             </div>
             <div className="row mb-3">
