+++ src/pages/editionCv/Editlangue.js
@@ -4,8 +4,8 @@ import {NotificationManager} from 'react-notifications';
 
 export default function EditLangue({id}) {
     const [langue, setLangue] = useState('');
-    const [langage, setLangage] = useState('');
-    const [progresslangue, setProgresslangue] = useState('');
+    const [language, setlanguage] = useState('');
+    const [progresslang, setProgresslang] = useState('');
 
     useEffect(() => {
         axios.get(`langages/${id}`).then(resp => {
@@ -17,27 +17,29 @@ export default function EditLangue({id}) {
 
     const editInfo = () => {
         const formdata = new FormData;
-        if(langue){formdata.append('langue', langue)}
-        if(progresslangue){formdata.append('progresslangue', progresslangue)}
-        if(langue || progresslangue) {
-            axios.put(`langages/${id}`, formdata).then(resp => {
-                if(resp.status === 200) {
-                    NotificationManager.success('Modifier avec succée', 'Modification valdataer', 4000)
-                }
-            }).catch(error => console.log(error))
-        }else {
-            NotificationManager.warning('Aucune modification enregistrer', 'Modification non valider', 4000)
-        }
+        if(language){formdata.append('langue', language)}
+        if(progresslang){formdata.append('progresslangue', progresslang)}
+        axios.put(`langages/${id}`, formdata).then(resp => {
+            if(resp.status === 200) {
+                NotificationManager.success('Modifier avec succée', 'Modification valdataer', 4000)
+            }
+        }).catch(error => console.log(error))
+        
     }
 
     const deleteLangue = () => {
-        axios.delete(`langages/${id}`).then(resp => {
-            if(resp.status === 204) {
-                NotificationManager.success('Vous avec supprimer une langue', 'Suprimer avec succèes', 4000)
-            }
-        }).catch(error => console.log(error))
+        if(window.confirm("Vous êtes Sûr ?") === true){
+            axios.delete(`langages/${id}`).then(resp => {
+                if(resp.status === 204) {
+                    NotificationManager.success('Vous avec supprimer une langue', 'Suprimer avec succèes', 4000)
+                }
+            }).catch(error => console.log(error))
+        }
     }
-
+    const niveau = [];
+    for(let numbers = 10; numbers <= 100; numbers += 10) {
+        niveau.push(`${numbers}`)
+    }
     return (
         <>
             <div className="row mb-3">
@@ -47,7 +49,7 @@ export default function EditLangue({id}) {
                 <div className="col-sm-9 text-secondary">
                     <input 
                         type="text" className="form-control" defaultValue={langue.langue}
-                        onChange={(e) => setLangage(e.target.value)}
+                        onChange={(e) => setlanguage(e.target.value)}
                     />
                 </div>
             </div>
@@ -56,10 +58,18 @@ export default function EditLangue({id}) {
                     <h6 className="mb-0">Niveau </h6>
                 </div>
                 <div className="col-sm-9 text-secondary input-group">
-                    <input 
-                        type="text" className="form-control" defaultValue={langue.progresslangue}
-                        onChange={(e) => setProgresslangue(e.target.value)}
-                    />
+                    <select 
+                        type="select" className="form-control"
+                        onChange={(e) => setProgresslang(e.target.value)}
+                    >
+                        <option>{langue.progresslangue}</option>
+                        {niveau && niveau.map((niv) => (
+                            <>
+                                <option value={niv}>{niv}</option>
+                            </>
+                        ))}
+
+                    </select>
                     <span class="input-group-text">%</span>
                 </div>
             </div>
@@ -68,7 +78,7 @@ export default function EditLangue({id}) {
                 <div className="col-sm-3"></div>
                 <div className="col-sm-9 text-secondary ">
                     <div className="btn-group" role="group" aria-label="Basic mixed styles example">
-                        <button title="Enregistrer les modifications" type="button" onClick={EditLangue} className="btn btn-primary">
+                        <button title="Enregistrer les modifications" type="button" onClick={editInfo} className="btn btn-primary">
                             <i class="bi bi-bookmark-fill"></i>&nbsp;
                             Enregistrer
                         </button>
